+++ dgraph/cmd/alpha/run.go
@@ -189,7 +189,7 @@ they form a Raft group and provide synchronous replication.
 	grpc.EnableTracing = false
 
 	flag.Bool("graphql_introspection", true, "Set to false for no GraphQL schema introspection")
-
+	flag.Bool("ludicrous_mode", false, "Run alpha in ludicrous mode")
 }
 
 func setupCustomTokenizers() {
@@ -459,8 +459,9 @@ func setupServer(closer *y.Closer) {
 	whitelist := func(h http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			if !handlerInit(w, r, map[string]bool{
-				http.MethodPost: true,
-				http.MethodGet:  true,
+				http.MethodPost:    true,
+				http.MethodGet:     true,
+				http.MethodOptions: true,
 			}) {
 				return
 			}
@@ -583,6 +584,7 @@ func run() {
 		SnapshotAfter:       Alpha.Conf.GetInt("snapshot_after"),
 		AbortOlderThan:      abortDur,
 		StartTime:           startTime,
+		LudicrousMode:       Alpha.Conf.GetBool("ludicrous_mode"),
 	}
 
 	setupCustomTokenizers()
@@ -591,8 +593,16 @@ func run() {
 	x.Config.QueryEdgeLimit = cast.ToUint64(Alpha.Conf.GetString("query_edge_limit"))
 	x.Config.NormalizeNodeLimit = cast.ToInt(Alpha.Conf.GetString("normalize_node_limit"))
 
-	x.PrintVersion()
+	x.InitSentry(enc.EeBuild)
+	defer x.FlushSentry()
+	x.ConfigureSentryScope("alpha")
+	x.WrapPanics()
 
+	// Simulate a Sentry exception or panic event as shown below.
+	// x.CaptureSentryException(errors.New("alpha exception"))
+	// x.Panic(errors.New("alpha manual panic will send 2 events"))
+
+	x.PrintVersion()
 	glog.Infof("x.Config: %+v", x.Config)
 	glog.Infof("x.WorkerConfig: %+v", x.WorkerConfig)
 	glog.Infof("worker.Config: %+v", worker.Config)