+++ ee/acl/acl_test.go
@@ -1010,7 +1010,7 @@ func TestQueryRemoveUnauthorizedPred(t *testing.T) {
 		{
 			`
 			{
-				me(func: has(name)) {
+				me(func: has(name), orderasc: name) {
 					name
 					age
 				}
@@ -1022,7 +1022,7 @@ func TestQueryRemoveUnauthorizedPred(t *testing.T) {
 		{
 			`
 			{
-				me(func: has(age)) {
+				me(func: has(age), orderasc: name) {
 					name
 					age
 				}
@@ -1035,22 +1035,20 @@ func TestQueryRemoveUnauthorizedPred(t *testing.T) {
 			`
 			{
 				me1(func: has(name), orderdesc: age) {
-					name
 					age
 				}
 				me2(func: has(name), orderasc: age) {
-					name
 					age
 				}
 			}
 			`,
-			`{"me1":[{"name":"RandomGuy"},{"name":"RandomGuy2"}],"me2":[{"name":"RandomGuy"},{"name":"RandomGuy2"}]}`,
+			`{"me1":[],"me2":[]}`,
 			`me1, me2 will have same results, can't order by <age> since it is unauthorized`,
 		},
 		{
 			`
 			{
-				me(func: has(name)) @groupby(age) {
+				me(func: has(name), orderasc: name) @groupby(age) {
 					count(name)
 				}
 			}
@@ -1061,7 +1059,7 @@ func TestQueryRemoveUnauthorizedPred(t *testing.T) {
 		{
 			`
 			{
-				me(func: has(name)) @filter(eq(nickname, "RG")) {
+				me(func: has(name), orderasc: name) @filter(eq(nickname, "RG")) {
 					name
 					age
 				}
@@ -1874,4 +1872,61 @@ func TestAddUpdateGroupWithDuplicateRules(t *testing.T) {
 
 	// cleanup
 	deleteGroup(t, grootJwt, groupName)
+}
+
+func TestAllowUIDAccess(t *testing.T) {
+	ctx, _ := context.WithTimeout(context.Background(), 100*time.Second)
+
+	dg, err := testutil.DgraphClientWithGroot(testutil.SockAddr)
+	require.NoError(t, err)
+
+	testutil.DropAll(t, dg)
+	op := api.Operation{Schema: `
+		name	 : string @index(exact) .
+	`}
+	require.NoError(t, dg.Alter(ctx, &op))
+	require.NoError(t, testutil.WaitForAlter(ctx, dg, op.Schema))
+
+	resetUser(t)
+	accessJwt, _, err := testutil.HttpLogin(&testutil.LoginParams{
+		Endpoint: adminEndpoint,
+		UserID:   "groot",
+		Passwd:   "password",
+	})
+	require.NoError(t, err, "login failed")
+	createGroup(t, accessJwt, devGroup)
+	addToGroup(t, accessJwt, userid, devGroup)
+
+	require.NoError(t, testutil.AssignUids(101))
+	mutation := &api.Mutation{
+		SetNquads: []byte(`
+			<100> <name> "100th User" .
+		`),
+		CommitNow: true,
+	}
+	_, err = dg.NewTxn().Mutate(ctx, mutation)
+	require.NoError(t, err)
+
+	// give read access of <name> to alice
+	addRulesToGroup(t, accessJwt, devGroup, []rule{{"name", Read.Code}})
+
+	userClient, err := testutil.DgraphClient(testutil.SockAddr)
+	require.NoError(t, err)
+	time.Sleep(6 * time.Second)
+
+	err = userClient.Login(ctx, userid, userpassword)
+	require.NoError(t, err)
+
+	uidQuery := `
+	{
+		me(func: uid(100)) {
+			uid
+			name
+		}
+	}
+	`
+
+	resp, err := userClient.NewReadOnlyTxn().Query(ctx, uidQuery)
+	require.Nil(t, err)
+	testutil.CompareJSON(t, `{"me":[{"name":"100th User", "uid": "0x64"}]}`, string(resp.GetJson()))
 }