+++ posting/index.go
@@ -24,11 +24,11 @@ import (
 	"io/ioutil"
 	"math"
 	"os"
-	"path/filepath"
 	"sync/atomic"
 	"time"
 
 	"github.com/golang/glog"
+	"github.com/pkg/errors"
 	ostats "go.opencensus.io/stats"
 	otrace "go.opencensus.io/trace"
 
@@ -40,7 +40,6 @@ import (
 	"github.com/dgraph-io/dgraph/tok"
 	"github.com/dgraph-io/dgraph/types"
 	"github.com/dgraph-io/dgraph/x"
-	"github.com/pkg/errors"
 )
 
 var emptyCountParams countParams
@@ -52,9 +51,9 @@ type indexMutationInfo struct {
 	op         pb.DirectedEdge_Op
 }
 
-// indexTokensforTokenizers return tokens, without the predicate prefix and
+// indexTokens return tokens, without the predicate prefix and
 // index rune, for specific tokenizers.
-func indexTokens(info *indexMutationInfo) ([]string, error) {
+func indexTokens(ctx context.Context, info *indexMutationInfo) ([]string, error) {
 	attr := info.edge.Attr
 	lang := info.edge.GetLang()
 
@@ -63,7 +62,7 @@ func indexTokens(info *indexMutationInfo) ([]string, error) {
 		return nil, errors.Errorf("Cannot index attribute %s of type object.", attr)
 	}
 
-	if !schema.State().IsIndexed(attr) {
+	if !schema.State().IsIndexed(ctx, attr) {
 		return nil, errors.Errorf("Attribute %s is not indexed.", attr)
 	}
 	sv, err := types.Convert(info.val, schemaType)
@@ -87,14 +86,13 @@ func indexTokens(info *indexMutationInfo) ([]string, error) {
 // TODO - See if we need to pass op as argument as t should already have Op.
 func (txn *Txn) addIndexMutations(ctx context.Context, info *indexMutationInfo) error {
 	if info.tokenizers == nil {
-		info.tokenizers = schema.State().Tokenizer(info.edge.Attr)
+		info.tokenizers = schema.State().Tokenizer(ctx, info.edge.Attr)
 	}
 
 	attr := info.edge.Attr
 	uid := info.edge.Entity
 	x.AssertTrue(uid != 0)
-	tokens, err := indexTokens(info)
-
+	tokens, err := indexTokens(ctx, info)
 	if err != nil {
 		// This data is not indexable
 		return err
@@ -115,10 +113,8 @@ func (txn *Txn) addIndexMutations(ctx context.Context, info *indexMutationInfo) 
 	return nil
 }
 
-func (txn *Txn) addIndexMutation(ctx context.Context, edge *pb.DirectedEdge,
-	token string) error {
+func (txn *Txn) addIndexMutation(ctx context.Context, edge *pb.DirectedEdge, token string) error {
 	key := x.IndexKey(edge.Attr, token)
-
 	plist, err := txn.cache.GetFromDelta(key)
 	if err != nil {
 		return err
@@ -198,7 +194,7 @@ func (txn *Txn) addReverseMutation(ctx context.Context, t *pb.DirectedEdge) erro
 
 func (txn *Txn) addReverseAndCountMutation(ctx context.Context, t *pb.DirectedEdge) error {
 	key := x.ReverseKey(t.Attr, t.ValueId)
-	hasCountIndex := schema.State().HasCount(t.Attr)
+	hasCountIndex := schema.State().HasCount(ctx, t.Attr)
 
 	var getFn func(key []byte) (*List, error)
 	if hasCountIndex {
@@ -238,11 +234,10 @@ func (txn *Txn) addReverseAndCountMutation(ctx context.Context, t *pb.DirectedEd
 	return nil
 }
 
-func (l *List) handleDeleteAll(ctx context.Context, edge *pb.DirectedEdge,
-	txn *Txn) error {
-	isReversed := schema.State().IsReversed(edge.Attr)
-	isIndexed := schema.State().IsIndexed(edge.Attr)
-	hasCount := schema.State().HasCount(edge.Attr)
+func (l *List) handleDeleteAll(ctx context.Context, edge *pb.DirectedEdge, txn *Txn) error {
+	isReversed := schema.State().IsReversed(ctx, edge.Attr)
+	isIndexed := schema.State().IsIndexed(ctx, edge.Attr)
+	hasCount := schema.State().HasCount(ctx, edge.Attr)
 	delEdge := &pb.DirectedEdge{
 		Attr:   edge.Attr,
 		Op:     edge.Op,
@@ -264,7 +259,7 @@ func (l *List) handleDeleteAll(ctx context.Context, edge *pb.DirectedEdge,
 				Value: p.Value,
 			}
 			return txn.addIndexMutations(ctx, &indexMutationInfo{
-				tokenizers: schema.State().Tokenizer(edge.Attr),
+				tokenizers: schema.State().Tokenizer(ctx, edge.Attr),
 				edge:       edge,
 				val:        val,
 				op:         pb.DirectedEdge_DEL,
@@ -404,8 +399,7 @@ func (txn *Txn) addMutationHelper(ctx context.Context, l *List, doUpdateIndex bo
 
 // AddMutationWithIndex is addMutation with support for indexing. It also
 // supports reverse edges.
-func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge,
-	txn *Txn) error {
+func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge, txn *Txn) error {
 	if len(edge.Attr) == 0 {
 		return errors.Errorf("Predicate cannot be empty for edge with subject: [%v], object: [%v]"+
 			" and value: [%v]", edge.Entity, edge.ValueId, edge.Value)
@@ -415,8 +409,8 @@ func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge,
 		return l.handleDeleteAll(ctx, edge, txn)
 	}
 
-	doUpdateIndex := pstore != nil && schema.State().IsIndexed(edge.Attr)
-	hasCountIndex := schema.State().HasCount(edge.Attr)
+	doUpdateIndex := pstore != nil && schema.State().IsIndexed(ctx, edge.Attr)
+	hasCountIndex := schema.State().HasCount(ctx, edge.Attr)
 	val, found, cp, err := txn.addMutationHelper(ctx, l, doUpdateIndex, hasCountIndex, edge)
 	if err != nil {
 		return err
@@ -431,7 +425,7 @@ func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge,
 		// Exact matches.
 		if found && val.Value != nil {
 			if err := txn.addIndexMutations(ctx, &indexMutationInfo{
-				tokenizers: schema.State().Tokenizer(edge.Attr),
+				tokenizers: schema.State().Tokenizer(ctx, edge.Attr),
 				edge:       edge,
 				val:        val,
 				op:         pb.DirectedEdge_DEL,
@@ -445,7 +439,7 @@ func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge,
 				Value: edge.Value,
 			}
 			if err := txn.addIndexMutations(ctx, &indexMutationInfo{
-				tokenizers: schema.State().Tokenizer(edge.Attr),
+				tokenizers: schema.State().Tokenizer(ctx, edge.Attr),
 				edge:       edge,
 				val:        val,
 				op:         pb.DirectedEdge_SET,
@@ -456,7 +450,7 @@ func (l *List) AddMutationWithIndex(ctx context.Context, edge *pb.DirectedEdge,
 	}
 	// Add reverse mutation irrespective of hasMutated, server crash can happen after
 	// mutation is synced and before reverse edge is synced
-	if (pstore != nil) && (edge.ValueId != 0) && schema.State().IsReversed(edge.Attr) {
+	if (pstore != nil) && (edge.ValueId != 0) && schema.State().IsReversed(ctx, edge.Attr) {
 		if err := txn.addReverseAndCountMutation(ctx, edge); err != nil {
 			return err
 		}
@@ -485,7 +479,7 @@ func deleteTokensFor(attr, tokenizerName string, hasLang bool) error {
 	// Also delete all the parts of any list that has been split into multiple parts.
 	// Such keys have a different prefix (the last byte is set to 1).
 	prefix = pk.IndexPrefix()
-	prefix[len(prefix)-1] = x.ByteSplit
+	prefix[0] = x.ByteSplit
 	prefix = append(prefix, tokenizer.Identifier())
 	return pstore.DropPrefix(prefix)
 }
@@ -500,7 +494,7 @@ func deleteReverseEdges(attr string) error {
 	// Also delete all the parts of any list that has been split into multiple parts.
 	// Such keys have a different prefix (the last byte is set to 1).
 	prefix = pk.ReversePrefix()
-	prefix[len(prefix)-1] = x.ByteSplit
+	prefix[0] = x.ByteSplit
 
 	return pstore.DropPrefix(prefix)
 }
@@ -517,13 +511,14 @@ func deleteCountIndex(attr string) error {
 	// Also delete all the parts of any list that has been split into multiple parts.
 	// Such keys have a different prefix (the last byte is set to 1).
 	prefix := pk.CountPrefix(false)
-	prefix[len(prefix)-1] = x.ByteSplit
+	prefix[0] = x.ByteSplit
 	if err := pstore.DropPrefix(prefix); err != nil {
 		return err
 	}
 
+	// Delete parts for count-reverse index.
 	prefix = pk.CountPrefix(true)
-	prefix[len(prefix)-1] = x.ByteSplit
+	prefix[0] = x.ByteSplit
 	return pstore.DropPrefix(prefix)
 }
 
@@ -539,22 +534,14 @@ type rebuilder struct {
 }
 
 func (r *rebuilder) Run(ctx context.Context) error {
-	// All the temp indexes go into the following directory. We delete the whole
-	// directory after the indexing step is complete. This deletes any other temp
-	// indexes that may have been left around in case defer wasn't executed.
-	// TODO(Aman): If users are not happy, we could add a flag to choose this dir.
-	tmpParentDir := filepath.Join(os.TempDir(), "dgraph_index")
-
 	// We write the index in a temporary badger first and then,
 	// merge entries before writing them to p directory.
-	if err := os.MkdirAll(tmpParentDir, os.ModePerm); err != nil {
-		return errors.Wrap(err, "error creating in temp dir for reindexing")
-	}
-	tmpIndexDir, err := ioutil.TempDir(tmpParentDir, "")
+	// TODO(Aman): If users are not happy, we could add a flag to choose this dir.
+	tmpIndexDir, err := ioutil.TempDir("", "dgraph_index_")
 	if err != nil {
 		return errors.Wrap(err, "error creating temp dir for reindexing")
 	}
-	defer os.RemoveAll(tmpParentDir)
+	defer os.RemoveAll(tmpIndexDir)
 	glog.V(1).Infof("Rebuilding indexes using the temp folder %s\n", tmpIndexDir)
 
 	dbOpts := badger.DefaultOptions(tmpIndexDir).
@@ -576,11 +563,11 @@ func (r *rebuilder) Run(ctx context.Context) error {
 		"Rebuilding index for predicate %s: Starting process. StartTs=%d. Prefix=\n%s\n",
 		r.attr, r.startTs, hex.Dump(r.prefix))
 
-	// Counter is used here to ensure that all keys are commited at different timestamp.
+	// Counter is used here to ensure that all keys are committed at different timestamp.
 	// We set it to 1 in case there are no keys found and NewStreamAt is called with ts=0.
 	var counter uint64 = 1
 
-	// Todo(Aman): Replace TxnWriter with WriteBatch. While we do that we should ensure that
+	// TODO(Aman): Replace TxnWriter with WriteBatch. While we do that we should ensure that
 	// WriteBatch has a mechanism for throttling. Also, find other places where TxnWriter
 	// could be replaced with WriteBatch in the code
 	tmpWriter := NewTxnWriter(tmpDB)
@@ -714,12 +701,65 @@ const (
 	indexRebuild         = iota // Index should be deleted and rebuilt.
 )
 
-// Run rebuilds all indices that need it.
-func (rb *IndexRebuild) Run(ctx context.Context) error {
-	if err := rebuildListType(ctx, rb); err != nil {
+// GetQuerySchema returns the schema that can be served while indexes are getting built.
+// Query schema is defined as current schema minus tokens to delete from current schema.
+func (rb *IndexRebuild) GetQuerySchema() *pb.SchemaUpdate {
+	// Copy the current schema.
+	querySchema := *rb.CurrentSchema
+	info := rb.needsTokIndexRebuild()
+
+	// Compute old.Tokenizer minus info.tokenizersToDelete.
+	interimTokenizers := make([]string, 0)
+	for _, t1 := range rb.OldSchema.Tokenizer {
+		found := false
+		for _, t2 := range info.tokenizersToDelete {
+			if t1 == t2 {
+				found = true
+				break
+			}
+		}
+		if !found {
+			interimTokenizers = append(interimTokenizers, t1)
+		}
+	}
+	querySchema.Tokenizer = interimTokenizers
+
+	if rb.needsCountIndexRebuild() == indexRebuild {
+		querySchema.Count = false
+	}
+	if rb.needsReverseEdgesRebuild() == indexRebuild {
+		querySchema.Directive = pb.SchemaUpdate_NONE
+	}
+	return &querySchema
+}
+
+// DropIndexes drops the indexes that need to be rebuilt.
+func (rb *IndexRebuild) DropIndexes(ctx context.Context) error {
+	if err := dropTokIndexes(ctx, rb); err != nil {
 		return err
 	}
-	if err := rebuildIndex(ctx, rb); err != nil {
+	if err := dropReverseEdges(ctx, rb); err != nil {
+		return err
+	}
+	return dropCountIndex(ctx, rb)
+}
+
+// BuildData updates data.
+func (rb *IndexRebuild) BuildData(ctx context.Context) error {
+	return rebuildListType(ctx, rb)
+}
+
+// NeedIndexRebuild returns true if any of the tokenizer, reverse
+// or count indexes need to be rebuilt.
+func (rb *IndexRebuild) NeedIndexRebuild() bool {
+	return rb.needsTokIndexRebuild().op == indexRebuild ||
+		rb.needsReverseEdgesRebuild() == indexRebuild ||
+		rb.needsCountIndexRebuild() == indexRebuild
+}
+
+// BuildIndexes builds indexes.
+func (rb *IndexRebuild) BuildIndexes(ctx context.Context) error {
+	if err := rebuildTokIndex(ctx, rb); err != nil {
 		return err
 	}
 	if err := rebuildReverseEdges(ctx, rb); err != nil {
@@ -734,7 +774,7 @@ type indexRebuildInfo struct {
 	tokenizersToRebuild []string
 }
 
-func (rb *IndexRebuild) needsIndexRebuild() indexRebuildInfo {
+func (rb *IndexRebuild) needsTokIndexRebuild() indexRebuildInfo {
 	x.AssertTruef(rb.CurrentSchema != nil, "Current schema cannot be nil.")
 
 	// If the old schema is nil, we can treat it as an empty schema. Copy it
@@ -802,12 +842,8 @@ func (rb *IndexRebuild) needsIndexRebuild() indexRebuildInfo {
 	}
 }
 
-// rebuildIndex rebuilds index for a given attribute.
-// We commit mutations with startTs and ignore the errors.
-func rebuildIndex(ctx context.Context, rb *IndexRebuild) error {
-	// Exit early if indices do not need to be rebuilt.
-	rebuildInfo := rb.needsIndexRebuild()
-
+func dropTokIndexes(ctx context.Context, rb *IndexRebuild) error {
+	rebuildInfo := rb.needsTokIndexRebuild()
 	if rebuildInfo.op == indexNoop {
 		return nil
 	}
@@ -826,17 +862,7 @@ func rebuildIndex(ctx context.Context, rb *IndexRebuild) error {
 		}
 	}
 
-	// Exit early if the index only need to be deleted and not rebuilt.
-	if rebuildInfo.op == indexDelete {
-		return nil
-	}
-
-	// Exit early if there are no tokenizers to rebuild.
-	if len(rebuildInfo.tokenizersToRebuild) == 0 {
-		return nil
-	}
-
-	glog.Infof("Rebuilding index for attr %s and tokenizers %s", rb.Attr,
+	glog.Infof("Deleting index for attr %s and tokenizers %s", rb.Attr,
 		rebuildInfo.tokenizersToRebuild)
 	// Before rebuilding, the existing index needs to be deleted.
 	for _, tokenizer := range rebuildInfo.tokenizersToRebuild {
@@ -851,6 +877,24 @@ func rebuildIndex(ctx context.Context, rb *IndexRebuild) error {
 		}
 	}
 
+	return nil
+}
+
+// rebuildTokIndex rebuilds index for a given attribute.
+// We commit mutations with startTs and ignore the errors.
+func rebuildTokIndex(ctx context.Context, rb *IndexRebuild) error {
+	rebuildInfo := rb.needsTokIndexRebuild()
+	if rebuildInfo.op != indexRebuild {
+		return nil
+	}
+
+	// Exit early if there are no tokenizers to rebuild.
+	if len(rebuildInfo.tokenizersToRebuild) == 0 {
+		return nil
+	}
+
+	glog.Infof("Rebuilding index for attr %s and tokenizers %s", rb.Attr,
+		rebuildInfo.tokenizersToRebuild)
 	tokenizers, err := tok.GetTokenizers(rebuildInfo.tokenizersToRebuild)
 	if err != nil {
 		return err
@@ -912,8 +956,8 @@ func (rb *IndexRebuild) needsCountIndexRebuild() indexOp {
 	return indexRebuild
 }
 
-// rebuildCountIndex rebuilds the count index for a given attribute.
-func rebuildCountIndex(ctx context.Context, rb *IndexRebuild) error {
+func dropCountIndex(ctx context.Context, rb *IndexRebuild) error {
+	// Exit early if indices do not need to be rebuilt.
 	op := rb.needsCountIndexRebuild()
 	if op == indexNoop {
 		return nil
@@ -924,8 +968,13 @@ func rebuildCountIndex(ctx context.Context, rb *IndexRebuild) error {
 		return err
 	}
 
-	// Exit early if attribute is index only needed to be deleted.
-	if op == indexDelete {
+	return nil
+}
+
+// rebuildCountIndex rebuilds the count index for a given attribute.
+func rebuildCountIndex(ctx context.Context, rb *IndexRebuild) error {
+	op := rb.needsCountIndexRebuild()
+	if op != indexRebuild {
 		return nil
 	}
 
@@ -996,20 +1045,20 @@ func (rb *IndexRebuild) needsReverseEdgesRebuild() indexOp {
 	return indexDelete
 }
 
-// rebuildReverseEdges rebuilds the reverse edges for a given attribute.
-func rebuildReverseEdges(ctx context.Context, rb *IndexRebuild) error {
+func dropReverseEdges(ctx context.Context, rb *IndexRebuild) error {
 	op := rb.needsReverseEdgesRebuild()
 	if op == indexNoop {
 		return nil
 	}
 
 	glog.Infof("Deleting reverse index for %s", rb.Attr)
-	if err := deleteReverseEdges(rb.Attr); err != nil {
-		return err
-	}
+	return deleteReverseEdges(rb.Attr)
+}
 
-	// Exit early if index only needed to be deleted.
-	if op == indexDelete {
+// rebuildReverseEdges rebuilds the reverse edges for a given attribute.
+func rebuildReverseEdges(ctx context.Context, rb *IndexRebuild) error {
+	op := rb.needsReverseEdgesRebuild()
+	if op != indexRebuild {
 		return nil
 	}